refactor(dashboard): extract runtime + keeper-cost domains (A1 phase 2 cont.)
- dashboard-runtime.ts: runtime providers/model metrics + runtime.toml config
  (providers 143-565 + toml 블록 통합). runtimeLaneLabel moved here from
  runtime, then duplicated into keeper-cost where the cost-matrix decoder
  depends on it (dashboard-runtime drop).
- dashboard-keeper-cost.ts: keeper cost metrics/decisions + cost-latency
  (metrics 162-339 + latency 블록 통합) + runtimeLaneLabel copy for decoder.
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: parseRuntimeDefaultsResponse, DEFAULT_WINDOW_MINUTES_24H,
  DashboardFeedMetadata, decodeDashboardFeedMetadata (all moved to domain files)
- typecheck green (tsc --noEmit)

dashboard.ts 3067 -> 2339 lines after runtime + keeper-cost.
